Stub out event history view pages
