+++ src/lib/components/select/filter-select.svelte
@@ -4,7 +4,7 @@
   import { updateQueryParameters } from '$lib/utilities/update-query-parameters';
   import Select from './select.svelte';
 
-  export let label: string;
+  export let label: string = null;
   export let value: string;
   export let parameter: string = null;
