+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/history/compact/index.svelte
@@ -1,0 +1,1 @@
+<p>Compact View</p>
