+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/history/full/index.svelte
@@ -1,0 +1,1 @@
+<p>Full View</p>
